refactor: move markdown to how to use folder
