+++ src/locales/how-to-use/markdowns/index.ts
